requirements.txt
cryptography~=43.0.0
pycryptodome
requests~=2.32.3
platform
webbrowser
smtplib
